refactor: コンポーネントの配置を変更
