+++ src/app/activities/page.tsx
@@ -1,4 +1,4 @@
-import { DescriptionWithImage } from '../about/description-with-image';
+import { DescriptionWithImage } from '@/features/about/components/description-with-image';
 
 export default function SeminarPage() {
   return (
