+++ src/features/about/components/description-with-image.tsx
